+++ tests/write_test.rs
@@ -38,12 +38,8 @@ fn create_and_encode_image() -> Result<()> {
     // Check result of encoding by decode it
     let context = HeifContext::read_from_bytes(&buf)?;
     let handle = context.primary_image_handle()?;
-
-    // TODO: libheif 1.10.0 has an error with saving images.
-    //       Saved image don't has correct information about width and height.
-    //       https://github.com/strukturag/libheif/issues/405
-    // assert_eq!(handle.width(), width);
-    // assert_eq!(handle.height(), height);
+    assert_eq!(handle.width(), width);
+    assert_eq!(handle.height(), height);
 
     // Decode the image
     let image = handle.decode(ColorSpace::Rgb(RgbChroma::Rgb), false)?;